% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{execute}
\alias{execute}
\title{Execute function for system calls}
\usage{
execute(x, outputfile = NA, intern = FALSE, quitOnError = FALSE)
}
\arguments{
\item{x}{Constructed command}

\item{outputfile}{Name for output file of command. Default = NA}

\item{intern}{To set the intern option of the internal system function. Defaults to FALSE}

\item{quitonError}{Decide to close R if an error occurs. Defaults to FALSE}
}
\description{
This function takes a constructed command that is to be run on the command line
and executes that command if a specified output file does not already exist.
}
\examples{
execute(constructed.command, output.file)
}
\keyword{command}
\keyword{execute}
\keyword{system}
